Minor Cleanup (spellEffects)
No reason to nullify value when we load it a new value straight after.
Even though it may be the easiest solution what we're doing here from a
programming standpoint, I don't blizzard did a 50/50 chance on this one.
Raising the comment basically so I won't forget to look into it more.
There are only 2 mount speeds in vanilla.